+++ pkg/config/resolver.go
@@ -52,16 +52,41 @@ type Resolver struct {
 }
 
 // APIServer resolver method
-func (r *Resolver) APIServer(synced func() bool) api.Server {
+func (r *Resolver) APIServer(ctx context.Context, synced func() bool) api.Server {
 	var logger *zap.Logger
 	if r.config.API.Logging {
 		logger, _ = r.Logger()
+	}
+
+	if r.config.API.BasicAuth.SecretRef != "" {
+		values, err := r.SecretClient().Get(ctx, r.config.API.BasicAuth.SecretRef)
+		if err != nil {
+			zap.L().Error("failed to load basic auth secret", zap.Error(err))
+		}
+
+		if values.Username != "" {
+			r.config.API.BasicAuth.Username = values.Username
+		}
+		if values.Password != "" {
+			r.config.API.BasicAuth.Password = values.Password
+		}
+	}
+
+	var auth *api.BasicAuth
+	if r.config.API.BasicAuth.Username != "" && r.config.API.BasicAuth.Password != "" {
+		auth = &api.BasicAuth{
+			Username: r.config.API.BasicAuth.Username,
+			Password: r.config.API.BasicAuth.Password,
+		}
+
+		zap.L().Info("API BasicAuth enabled")
 	}
 
 	return api.NewServer(
 		r.TargetClients(),
 		r.config.API.Port,
 		logger,
+		auth,
 		synced,
 	)
 }