+++ pkg/config/resolver_test.go
@@ -292,7 +292,7 @@ func Test_ResolvePolicyStore(t *testing.T) {
 func Test_ResolveAPIServer(t *testing.T) {
 	resolver := config.NewResolver(&config.Config{}, &rest.Config{})
 
-	server := resolver.APIServer(func() bool { return true })
+	server := resolver.APIServer(context.Background(), func() bool { return true })
 	if server == nil {
 		t.Error("Error: Should return API Server")
 	}